requirements.txt
Flask
pandas
openpyxl
docxtpl
pywin32
docx2pdf
pyinstaller
lxml
pillow
jinja2
python-pptx
pywin32 comtypes